+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased
+* add ability to specify attributes and docs on each field individually by [cyqsimon](https://github.com/cyqsimon)
+* add an `append` mode to `doc` argument by [cyqsimon](https://github.com/cyqsimon)
+
 ## 0.3.0
 * update to syn v2.0, bumping minimum rustc version to 1.56.0
 * add feature list and attribute order note to documentation